+++ lua/kitty-scrollback/api.lua
@@ -80,10 +80,14 @@ M.generate_kittens = function(all, generate_modes)
     target_gen_modes[gen_mode] = true
   end
 
-  local kitty_scrollback_kitten =
-    vim.api.nvim_get_runtime_file('python/kitty_scrollback_nvim.py', false)[1]
-  local example_path =
-    vim.api.nvim_get_runtime_file('lua/kitty-scrollback/configs/example.lua', false)[1]
+  local kitty_scrollback_kitten = vim.fn.fnamemodify(
+    vim.api.nvim_get_runtime_file('python/kitty_scrollback_nvim.py', false)[1],
+    ':p'
+  )
+  local example_path = vim.fn.fnamemodify(
+    vim.api.nvim_get_runtime_file('lua/kitty-scrollback/configs/example.lua', false)[1],
+    ':p'
+  )
 
   local action_alias = 'kitty_scrollback_nvim'
   local alias_config = {
@@ -181,8 +185,10 @@ M.generate_kittens = function(all, generate_modes)
 end
 
 M.checkhealth = function()
-  local kitty_scrollback_kitten =
-    vim.api.nvim_get_runtime_file('python/kitty_scrollback_nvim.py', false)[1]
+  local kitty_scrollback_kitten = vim.fn.fnamemodify(
+    vim.api.nvim_get_runtime_file('python/kitty_scrollback_nvim.py', false)[1],
+    ':p'
+  )
   -- NOTE(#58): nvim v0.9 support
   -- setup backports for v0.9 because checkhealth can be called outside of standard setup flow
   if vim.fn.has('nvim-0.10') <= 0 then
@@ -191,7 +197,8 @@ M.checkhealth = function()
   if vim.fn.has('nvim-0.9') > 0 then
     vim
       .system({
-        'kitty',
+        -- fallback to 'kitty' because checkhealth can be called outside of standard setup flow
+        (p and p.kitty_data and p.kitty_data.kitty_path) and p.kitty_data.kitty_path or 'kitty',
         '@',
         'kitten',
         kitty_scrollback_kitten,