+++ lua/kitty-scrollback/launch.lua
@@ -45,6 +45,7 @@ local M = {}
 ---@field kitty_opts KsbKittyOpts relevant kitty configuration values
 ---@field kitty_config_dir string kitty configuration directory path
 ---@field kitty_version table kitty version
+---@field kitty_path string kitty executable path
 
 ---@class KsbPrivate
 ---@field orig_columns number
@@ -182,6 +183,7 @@ local function set_options()
   vim.o.termguicolors = true
 
   -- preferred optional opts
+  vim.opt.shortmess:append('I') -- no intro message
   vim.o.laststatus = 0
   vim.o.scrolloff = 0
   vim.o.cmdheight = 0
@@ -316,6 +318,8 @@ M.setup = function(kitty_data_str)
     end
   end
 
+  set_options()
+
   ksb_util.setup(p, opts)
   ksb_kitty_cmds.setup(p, opts)
   ksb_win.setup(p, opts)
@@ -323,12 +327,12 @@ M.setup = function(kitty_data_str)
   ksb_autocmds.setup(p, opts)
   ksb_api.setup(p, opts)
   ksb_keymaps.setup(p, opts)
+
   local ok = ksb_hl.setup(p, opts)
   if ok then
     ksb_hl.set_highlights()
     ksb_kitty_cmds.open_kitty_loading_window(ksb_hl.get_highlights_as_env()) -- must be after opts and set highlights
   end
-  set_options()
 
   if
     opts.callbacks
@@ -395,11 +399,10 @@ M.launch = function()
         local term_buf_name = vim.api.nvim_buf_get_name(p.bufid)
         term_buf_name = term_buf_name:gsub('^(term://.-:).*', '%1kitty-scrollback.nvim')
         vim.api.nvim_buf_set_name(p.bufid, term_buf_name)
-        vim.api.nvim_set_option_value(
-          'winhighlight',
-          'Visual:KittyScrollbackNvimVisual',
-          { win = 0 }
-        )
+        vim.api.nvim_set_option_value('winhighlight', 'Visual:KittyScrollbackNvimVisual', {
+          scope = 'local',
+          win = 0,
+        })
         vim.api.nvim_buf_delete(vim.fn.bufnr('#'), { force = true }) -- delete alt buffer after rename
 
         if opts.restore_options then