+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## [2.4.3](https://github.com/mikesmithgh/kitty-scrollback.nvim/compare/v2.4.2...v2.4.3) (2023-12-11)
+
+
+### Bug Fixes
+
+* add kitty fallback for checkhealth ([#98](https://github.com/mikesmithgh/kitty-scrollback.nvim/issues/98)) ([6c93db4](https://github.com/mikesmithgh/kitty-scrollback.nvim/commit/6c93db4f4a82001cd1c6f3ff014481426dfa8648))
+
+## [2.4.2](https://github.com/mikesmithgh/kitty-scrollback.nvim/compare/v2.4.1...v2.4.2) (2023-12-07)
+
+
+### Bug Fixes
+
+* read kitty executable from path ([#94](https://github.com/mikesmithgh/kitty-scrollback.nvim/issues/94)) ([acdc317](https://github.com/mikesmithgh/kitty-scrollback.nvim/commit/acdc3170f20252ba2a4c4bb114b698d85995ae70)), closes [#95](https://github.com/mikesmithgh/kitty-scrollback.nvim/issues/95)
+
+## [2.4.1](https://github.com/mikesmithgh/kitty-scrollback.nvim/compare/v2.4.0...v2.4.1) (2023-12-01)
+
+
+### Bug Fixes
+
+* update orig_columns during WinResized event ([#85](https://github.com/mikesmithgh/kitty-scrollback.nvim/issues/85)) ([88e2063](https://github.com/mikesmithgh/kitty-scrollback.nvim/commit/88e20631e55fc37a7926b597636897e058eaa783)), closes [#75](https://github.com/mikesmithgh/kitty-scrollback.nvim/issues/75)
+
+# [2.4.0](https://github.com/mikesmithgh/kitty-scrollback.nvim/compare/v2.3.1...v2.4.0) (2023-12-01)
+
+
+### Features
+
+* clear selection in kitty after paste ([#84](https://github.com/mikesmithgh/kitty-scrollback.nvim/issues/84)) ([7cc5814](https://github.com/mikesmithgh/kitty-scrollback.nvim/commit/7cc581409cbd6bd601e4a66f3b3a21af81c22bf5)), closes [#76](https://github.com/mikesmithgh/kitty-scrollback.nvim/issues/76)
+
 ## [2.3.1](https://github.com/mikesmithgh/kitty-scrollback.nvim/compare/v2.3.0...v2.3.1) (2023-11-27)